backend/requirements.txt
fastapi
google-generativeai
httpx
yfinance
pydantic
python-dotenv
uvicorn

# Web framework
fastapi
uvicorn

# HTTP requests (for FinanceHub proxy)
httpx

# Environment variables
python-dotenv

# AI model SDK
google-generativeai

# Finance data fallback
yfinance
pandas
numpy
